+++ vue.config.js
@@ -1,3 +1,4 @@
 module.exports = {
+    lintOnSave: false,
     css: { extract: false }
 }